Implement kv.js as a server-side cache
In the front end, the browser does most of the
client-side caching regarding images. In the
backend, the best bang-for-your-buck for
implementing a cache will be making requests
for the palette. Since the operation requires calls
to multiple packages, implementing a cache will
substantially reduce time needed to generate
the palette if we have seen the image before.

In this use-case, Redis and memcached may be an
overkill because all we need is an in-memory
key/value store that stores an array of strings with
the ability to expire the items. We will use `kv.js`
in this case which is a much lighter weight option